requirements.txt
accelerate==0.31.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
argcomplete==3.3.0
arrow==1.3.0
attrs==23.2.0
audioread==3.0.1
blinker==1.8.2
boto3==1.34.102
botocore==1.34.102
bravado==11.0.3
bravado-core==6.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
denoising-diffusion-pytorch==2.0.12
distlib==0.3.8
einops==0.8.0
ema-pytorch==0.5.1
filelock==3.14.0
Flask==3.0.3
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.5.0
iniconfig==2.0.0
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
neptune==1.10.3
nest-asyncio==1.6.0
networkx==3.3
nox==2024.4.15
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.10.7
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.1
plotly==5.24.1
plotly-resampler==0.10.0
pluggy==1.5.0
pooch==1.8.1
protobuf==5.27.2
psutil==5.9.8
ptwt==0.1.8
pyarrow==16.1.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-fid==0.3.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
ray==2.32.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
swagger-spec-validator==3.0.3
sympy==1.12
tenacity==9.0.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
tsdownsample==0.1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.26.2
webcolors==1.13
websocket-client==1.8.0
Werkzeug==3.0.4
zipp==3.20.2
